Moved template+controller into subroute
