+++ app/templates/templates.hbs
@@ -13,6 +13,3 @@
     </div>
   </div>
 </div>
-
-  <!-- You have {{templatePoints}} template points. -->
-  <!-- <button type="button" class="btn btn&#45;outline&#45;danger my&#45;2" onclick={{action "rebirth"}}>End Empire for {{game.empire.nextManaPoints}} mana points</button> -->
